Move classes into a module
